+++ src/main/kotlin/edu/uoc/hagendazs/macadamianut/application/event/event/service/EventService.kt
@@ -4,8 +4,14 @@ import edu.uoc.hagendazs.macadamianut.application.event.event.entrypoint.output.
 import edu.uoc.hagendazs.macadamianut.application.event.event.model.dataClass.DBEvent
 
 interface EventService {
-    fun createEvent(newEvent: DBEvent, categoryName: String?): EventResponse?
-    fun updateEvent(eventToUpdate: DBEvent): EventResponse?
+    fun createEvent(
+        newEvent: DBEvent,
+        categoryName: String?,
+        labelIds: Collection<String>
+    ): EventResponse?
+
+    fun updateEvent(eventToUpdate: DBEvent,
+                    labelIds: Collection<String>): EventResponse?
     fun findById(eventId: String): EventResponse?
     fun findEventsWithFilters(
         labels: Collection<String>,