+++ src/main/kotlin/edu/uoc/hagendazs/macadamianut/application/event/event/service/impl/EventServiceImpl.kt
@@ -8,9 +8,7 @@ import edu.uoc.hagendazs.macadamianut.application.event.event.service.EventServi
 import edu.uoc.hagendazs.macadamianut.application.event.event.service.exceptions.EventAlreadyExistsException
 import edu.uoc.hagendazs.macadamianut.application.event.event.service.exceptions.EventDoesNotExistException
 import edu.uoc.hagendazs.macadamianut.application.event.event.service.exceptions.UnableToCreateEventFromGivenData
-import edu.uoc.hagendazs.macadamianut.application.event.eventOrganizer.service.EventOrganizerService
-import edu.uoc.hagendazs.macadamianut.application.event.label.service.LabelService
-import edu.uoc.hagendazs.macadamianut.application.media.service.MediaService
+import edu.uoc.hagendazs.macadamianut.application.event.label.model.repo.LabelRepo
 import mu.KotlinLogging
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
@@ -24,12 +22,22 @@ class EventServiceImpl : EventService {
     @Autowired
     private lateinit var categoryRepo: CategoryRepo
 
+    @Autowired
+    private lateinit var labelRepo: LabelRepo
+
     private val logger = KotlinLogging.logger {}
 
-    override fun createEvent(newEvent: DBEvent, categoryName: String?): EventResponse? {
+    override fun createEvent(newEvent: DBEvent, categoryName: String?, labelIds: Collection<String>): EventResponse? {
         // check duplicate event names
         eventRepo.findByName(newEvent.name)?.let {
             throw EventAlreadyExistsException("${it.name} already exists in ")
+        }
+        val labelsFromDb = labelRepo.findLabelsWithId(labelIds).map { it.id }
+
+        if (labelsFromDb.size != labelIds.size) {
+            throw UnableToCreateEventFromGivenData(
+                "The following label ids do not exist: ${labelIds.minus(labelsFromDb.toSet())}"
+            )
         }
         // check that category exists
         val category = categoryRepo.findByName(categoryName) ?: run {
@@ -39,16 +47,16 @@ class EventServiceImpl : EventService {
             )
         }
 
-        val createdEvent = eventRepo.create(newEvent.copy(categoryId = category.id))
+        val createdEvent = eventRepo.create(newEvent.copy(categoryId = category.id), labelIds)
         logger.info { "Event created with name ${newEvent.name} and id $newEvent.id" }
         return createdEvent
     }
 
-    override fun updateEvent(eventToUpdate: DBEvent): EventResponse? {
+    override fun updateEvent(eventToUpdate: DBEvent, labelIds: Collection<String>): EventResponse? {
         eventRepo.findById(eventToUpdate.id) ?: run {
             EventDoesNotExistException("Event with id ${eventToUpdate.id} does not exist in this server")
         }
-        return eventRepo.update(eventToUpdate)
+        return eventRepo.update(eventToUpdate, labelIds)
     }
 
     override fun findById(eventId: String): EventResponse? {