+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # MACADAMIA NUT CHANGELOG
 
 
+##v0.0.7
+Changes:
+- New / Update event request requires the following fields:
+  -     val labelIds: Collection<String>,
+
+Fixes:
+- ...
+
 ##v0.0.6
 Changes:
 - Event Request and event data class now require eventUrl (String)